Add sketchable grid, grid density setter button which prompts user input
Added in this commit is a sketch pad consisting of nxm divs the number of which are set
by user input. The divs change color from white to black on mouseover if the left click is
being pressed or on click on its own. A toolbar with a set density button which when clicked
brings up a prompt asking for user input for pixel density of the sketch pad which runs the
UpdateGrid() function creating a grid with the desired density.